Restructuring the codebase. Not sure about the way to go though. Pretty sure i will introduce new module to separate gui & cli
