+++ src/main/cx.ksim.dicelock/cx/ksim/dicelock/core/DiceLockOptions.java
@@ -1,0 +1,15 @@
+package cx.ksim.dicelock.core;
+
+public record DiceLockOptions(
+		int wordCount, String wordList, 
+		String delimiter,boolean verbose,
+		boolean shouldCapitalize ) {
+	
+	DiceLockOptions() {
+		this(6,"/wordlist,txt");
+	}	
+	
+	DiceLockOptions(int wordCount, String wordList) {
+		this(wordCount, wordList, " ", false, true);
+	}
+}
