+++ src/main/cx.ksim.dicelock/cx/ksim/dicelock/gui/FXMLController.java
@@ -1,4 +1,4 @@
-package cx.ksim.dicelock.fx;
+package cx.ksim.dicelock.gui;
 
  
 import javafx.event.ActionEvent;
