+++ src/test/cx.ksim.dicelock/cx/ksim/dicelock/core/DiceLockTest.java
@@ -1,4 +1,4 @@
-package cx.ksim.dicelock;
+package cx.ksim.dicelock.core;
 
 import static org.junit.Assert.assertEquals;
 
@@ -6,12 +6,12 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
 class DiceLockTest {
-	private DiceLock dicelock = new DiceLock();
+	private DiceLockOptions options = new DiceLockOptions();
 	
 	@ParameterizedTest(name = "Should generate {0} words.")
 	@ValueSource(ints = { 3, 6, 9 })
 	void generateWords(int count) {
-		assertEquals( dicelock.generateRandomWords( count, "/wordlist.txt", " " ).split( " " ).length, count );
+		assertEquals( DiceLock.generateRandomWords( options ).split( " " ).length, count );
 	}
 
 }
